Renamed MyGravity to AttractiveForce and added more possibilities
git-svn-id: https://openlierox.svn.sourceforge.net/svnroot/openlierox@4617 34602234-ff1f-0410-a465-ea8f3b77ab7f

+++ src/common/PhysicsLX56.cpp
@@ -146,11 +146,11 @@ public:
 
 					if(x<0) {
 						clip |= 0x01;
-						worm->pos().x=( posx+4 + cClient->getMap()->GetWidth() ) % cClient->getMap()->GetWidth();
+						worm->pos().x= (float)(( posx+4 + cClient->getMap()->GetWidth() ) % cClient->getMap()->GetWidth());
 					}
 					else {
 						clip |= 0x02;
-						worm->pos().x=( posx-4 + cClient->getMap()->GetWidth() ) % cClient->getMap()->GetWidth();
+						worm->pos().x=(float)(( posx-4 + cClient->getMap()->GetWidth() ) % cClient->getMap()->GetWidth());
 					}
 
 					// Bounce
@@ -212,11 +212,11 @@ public:
 
 					if(y<0) {
 						clip |= 0x04;
-						worm->pos().y= ( posy+5 + cClient->getMap()->GetHeight() ) % cClient->getMap()->GetHeight();
+						worm->pos().y= (float)(( posy+5 + cClient->getMap()->GetHeight() ) % cClient->getMap()->GetHeight());
 					}
 					else {
 						clip |= 0x08;
-						worm->pos().y= ( posy-5 + cClient->getMap()->GetHeight() ) % cClient->getMap()->GetHeight();
+						worm->pos().y= (float)(( posy-5 + cClient->getMap()->GetHeight() ) % cClient->getMap()->GetHeight());
 					}
 
 					//if(y>3 && !jump) {
